+++ src/haddock/modules/sampling/rigidbody/__init__.py
@@ -117,12 +117,18 @@ class HaddockModule(BaseCNSModule):
             # Create a model for the expected output
             model = prepare_expected_pdb(combination, idx, ".", self.name)
             # Set additional attributes
-            model.restr_fname=str(ambig_fname)
+            model.restr_fname = str(ambig_fname)
             model.seed = seed
 
             self.output_models.append(model)
 
-            job = CNSJob(cns_input, log_fname, err_fname, envvars=self.envvars)
+            job = CNSJob(
+                cns_input,
+                log_fname,
+                err_fname,
+                envvars=self.envvars,
+                output_pdb_files=[model.file_name],
+            )
             jobs.append(job)
         return jobs