+++ src/haddock/modules/scoring/emscoring/__init__.py
@@ -79,7 +79,13 @@ class HaddockModule(CNSScoringModule):
 
             self.output_models.append(expected_pdb)
 
-            job = CNSJob(scoring_input, scoring_out, err_fname, envvars=self.envvars)
+            job = CNSJob(
+                scoring_input,
+                scoring_out,
+                err_fname,
+                envvars=self.envvars,
+                output_pdb_files=[expected_pdb.file_name],
+            )
 
             jobs.append(job)