+++ src/haddock/modules/topology/topocg/__init__.py
@@ -273,12 +273,28 @@ class HaddockModule(BaseCNSModule):
                 # Add new job to the pool
                 output_filename = Path(f"{model.stem}.{Format.CNS_OUTPUT}")
                 err_fname = f"{model.stem}.cnserr"
+                if shape_dic[i]:
+                    output_pdb = Path(f"{model.stem}.{Format.PDB}")
+                    output_topology = Path(f"{model.stem}.{Format.TOPOLOGY}")
+                else:
+                    output_pdb = Path(
+                        gen_cg_filename("", model.stem, force_field=force_field)
+                    )
+                    output_topology = Path(
+                        gen_cg_filename(
+                            "",
+                            model.stem,
+                            force_field=force_field,
+                            ext=Format.TOPOLOGY,
+                        )
+                    )
                 job = CNSJob(
                     topocg_input,
                     output_filename,
                     err_fname,
                     envvars=self.envvars,
                     cns_exec=cns_exec,
+                    output_files=[output_pdb, output_topology],
                 )
 
                 jobs.append(job)